Re-worked register read/write
The HBI interface uses word-aligned addresses and the
byte-enable bits.

So far, we had inconsist handling of the simulated and
'real' wrapper (the AXI bus expects byte addresses and
generates correct access using address and length).

The new HBIRead/HBIWrite procedure expects byte-addresses
and byte-enable bits that are 1111, 0111, 0011 or 0001 to
indicate the length.

This version passes first hardware tests...