Lan9254ESC.vhd: FIX: made HBI mux a separate state machine.
In some cases (program.don, reply.valid flags, ...) ESC work
must not be delayed across multiple cycles (when the HBI is
switched away from the ESC). This led to spurious reset of
'don' before it was seen and stretched reply.valid signals
which acknowledged unadvertently.

Together with a master patch (missing mutex lock) the EoE finally
works concurrently with a user app!