hdl/MicroUDPIPMux.vhd, hdl/MicroUDPTx.vhd: FIX condition to exit loopback state
It is not sufficient to look for the end of the stream in loopback
state - there may not be a stream at all (ARP-REP).

Instead, only ACK the txReq once it is completely done (including the
forwarding state) [MicroUDPTx.vhd] and let MicroUDPIPMux wait
for (txReq.valid and txAck) to exit loopback.

(loopback was never exited when there was no payload).