Support boot mailbox; variable-length SM2, SM3 + SM0
Remove the restriction that the SM lengths programmed by
the master must match the EEPROM configuration. This is not
strictly necessary (except for SM1 which is still limited to
a hardcoded length).

Made SM0's length also configurable so that a bigger
(albeit RX only) mailbox may be used in BOOT stage.

Support BOOT mailbox also in the tool.